Lots of stuff:
Handle most (all?) of the page style / numbering magic here so the documents
don't have to do it individually.

Revise the \bcode / \ecode stuff so that the {verbatim} environment handles
it right directly.  \bcode / \ecode will be completely removed from all files
(to be checked in momentarily).

Have the {verbatim} environment get the samples indented a bit; this
appearantly had been attempted in the old code, but didn't work because
paragraphs weren't indented.

Make all headers, from chapters on down to subparagraphs, have sans-serif
titles.

\setindexsubitem{}:  New macro.  Replaces \renewcommand{\indexsubitem{}(...)}
	everywhere.  This allows LaTeX2HTML to be made to work correctly for
	this.  That was near impossible with the old mechanisms.

For all {*desc} environments, make the name of the described thing bold as
well as monospaced.

{opcodedesc} environment:  Don't index the byte code names; that doesn't seem
	terribly useful, and there are a lot of them.

\var{}:  More magic to make sure that the size is right even if embedded in
	\file{} or some other macro that uses the sans-serif font in running
	text.

\bfcode{}:  New macro.  Makes the font \code{} and bold.  (Was unreasonable
	using old LaTeX 2.09.)

\file{}:  Adjust the size of the sans-serif font a little.

\email{}, \url{},  Make these use the same font as \file{}, but not the
	surrounding single-quotes.

Update many comments.

Lots of minor nits and a little cleanliness.