First cut at dynamic types.  Too much code moved around to describe the changes in detail.  Specification of what changed:
If you set ``__dynamic__ = 1'' is a class statement that creates a new
type, the new type is dynamic.  A new type is also dynamic if at least
one of its base types is dynamic and no __dynamic__ flag is set in the
class statement.  By setting ``__dynamic__ = 0'' in the class
statement you can force a static type even if some bases are dynamic.

For a dynamic type t:

- t.__dict__ is t.__defined__      # they are the same object

- type(t.__dict__) is dictionary   # not a proxy

- t.<name> = <value> is equivalent to t.__dict__['<name>'] = <value>

- attribute lookup for dynamic types and their instances searches the
  __defined__ dict of the type and its base types, in MRO sequence
  (even if the base class is static)