Aargh!
During an innocent benchmark test I found out that a simple getattr
operation on an instance of a static type was 20x slower than one on
an instance of a dynamic type.  It took an hour of tracing in the
debugger to realize that the call override_slots(type, type->tp_dict)
at the end of type_init() was wrong, and should really be
override_slots(type, type->tp_defined).  The tp_getattro slot had
turned into a wrapper for __getattr__, which in turn was a wrapper
around the base class's tp_getattro slot.

(The solution isn't quite right, because slots are now inherited from
the "strongest base" only, while they should be inherited from all
bases; but that's a separate issue; it's no worse than before, and the
speed glitch is gone.)