Total rewrite of the descriptor object implementation, to get rid of the discriminated union.  The ultimate goal here was to make the definition of PyDescr_IsData() more reasonable: it not just tests whether the type has a __set__ method (i.e. a tp_descr_set slot).  If it does, it's a data descriptor, otherwise it's not.  This is important when an attribute is found in both the __dict__ of the class and the __dict__ of the instance: when the descriptor in the class is a data descriptor, it has precedence, but when it is a non-data (e.g. method) desscriptor, the instance __dict__ takes precedence.
Previously, this function had to have special dispensation to look
inside descriptor objects -- this is no longer necessary.

Note that the presence of a __set__ method doesn't mean that the
attribute can actually be assigned to: this is also used to make
attributes deliberately read-only, by making the __set__ method always
raise an exception.  (Q: should it raise AttributeError or TypeError?
Traditionally, this has raised TypeError!)