Policy change for rich comparison: if the second operand's type is a subtype of the first operand's type, try its comparison *before* trying the first operand.
This removes the need for hacks in string and complex rich comparisons
that test whether the operands have the proper comparison slot;
instead, we can use a straightforward typecheck.

This solves some issues related to make __dynamic__ the default: the
test failed if the comparison slot was slot_tp_richcompare.

It also solves the situation where a derived class overrides a base
class's rich comparison but wants to call the base class's rich
comparison implementation:

  class C(complex):
    def __eq__(self, other):
      print "Comparing", self, "to", other
      return complex.__eq__(self, other)

Here the base class implementation would refuse the explicit call
because of the test for a specific slot function.