Updated NEWS for version 2.2beta2.
Note that I added (guessed at ;) some items for the Mac platform.
Jack should proofread and make any necessary corrections in both the
branch (when he takes it over for the Mac release), and in the trunk
so that it'll be updated for 2.2rc1.  Corrections won't make it into
the Windows or source releases for 2.2b2, so that'll have to be okay.