A somewhat yucky but I think necessary change: introduce a separate date class.  This implements all the date shuffling; the datetime class then subclasses this and adds more precision and tzoffset fiddling.
It's yucky because there's a fair amount of code duplication, alas --
this can probably be reduced by thinking more about which interfaces
are needed.

It's also necessary because the more I think about it, the more we
need a succession of types:

- date -- for calendrical computations
- xxxx -- for date+time computations without timezone awareness
- yyyy -- local-timezone-and-DST-aware date+time computations
- zzzz -- UTC date+time computations
- datetime -- generalized timezone-and-DST-aware date+time computations

I'll try to clean this up some more after my shower.