Move VisitorBase from asdl_c to asdl.
Add Check visitor in asdl that looks for errors.  So far the only
error checking is to verify that the same constructor name isn't used
more than once (even in different types).  It's not sensible to reuse
names, because you couldn't tell which type was implied from the
constructor alone.