Lots of hacks great and small to the main() program, but I didn't touch the guts of the algorithm.
Added a module docstring/usage message.

Added a bunch of switches to train the system on an mbox of known good
and known spam messages (using PortableUnixMailbox only for now).
Uses the email package but does not decoding of message bodies.  Also,
allows you to specify a file for pickling the training data, and for
setting a threshold, above which messages get an X-Bayes-Score
header.  Also output messages (marked and unmarked) to an output file
for retraining.

Print some statistics at the end.