spamprob():  Commented some subtleties.
clearjunk():  Undid Guido's attempt to space-optimize this.  The problem
is that you can't delete entries from a dict that's being crawled over
by .iteritems(), which is why I (I suddenly recall) materialized a
list of words to be deleted the first time I wrote this.  It's a lot
better to materialize a list of to-be-deleted words than to materialize
the entire database in a dict.items() list.