Allow command line specification of tokenize functions     run w/ -t flag to override default tokenize function     run w/ -H flag to see list of tokenize functions
When adding a new tokenizer, make docstring a short description and add a
key/value pair to the tokenizers dict.  The key is what the user specifies.
The value is a tokenize function.

Added two new tokenizers - tokenize_wordpairs_foldcase and
tokenize_words_and_pairs.  It's not obvious that either is better than any
of the preexisting functions.

Should probably add info to the pickle which indicates the tokenizing
function used to build it.  This could then be the default for spam
detection runs.

Next step is to drive this with spam/non-spam corpora, selecting each of the
various tokenizer functions, and presenting the results in tabular form.