Gave the timedelta class .min, .max, and .resolution attributes, like the other classes here.  Added a test case.  .min and .max don't currently make sense for the Python implementation, since that's happy to use unbounded longs for the # of days.
Sometime last week I silently introduced the limitation that a timedelta
object must have a (after normalization) number of days expressible in
no more than 9 decimal digits (abs(timedeltaobj.days) < a billion).  This
was to make the C implementation feasible with finite pain.  It's a curious
consequence that -timedelta.max overflows, since we constrain the internal
microsecond and second members to be >= 0.