Prepare to implement the time type.  This is mostly massive macro renaming, since e.g. the old GET_HOUR will become too ambiguous to tolerate.  That in turn triggered another round of long-line folding.
The new obj_time.c is just a copy of obj_datetime.c, and isn't yet
compiled.  Added a struct decl and macros for the time type, so far
unreferenced.

Moved the struct decl for timedelta into the .h file, since it didn't
make sense that it was the only one unexposed.