Mounds of changes to implement the time type in C.  Moved all the time tests from test_datetime.py into test_both.py.  More tests are needed (e.g., for pickling of time objects -- this is implemented but not yet tested).
The Python implementation had a fancy "pretty" __str__ for time and timetz.
I don't know whether or not that's a good idea, as people argue endlessly
over what "pretty" means.  Since all the other types in this module with
an isoformat() method made __str__ a synonym, that's what I did here too,
and changed the Python implementation to match.  I'm not married to this,
I just don't want to endure arguments about this now.

Pretty str() for all types or pretty str() for none is what I want to see,
the latter is quickest to implement, and if pretty is desired that's not
critical to get done before tha alpha release.