Tried to implement timetz pickling.  This doesn't work in the C implementation if timetz.tzinfo is non-None; don't know why; it works in the Python implementation; the C part of the test is disabled for now.
Fixed a subtle bug in time.__hash__, only triggered if a subclass had
its own __getstate__ and ended up invoking time.__hash__.

Added some more error checks.

There's a new leak when running the tests on the C implementation;
don't know why; it's *not* in the timetz tests, so that makes it a
bit of a mystery.