Added helper method call randlist() to be used in place of the method    formerly included in the C code.
Revised the reference implementation test.  Now uses seeds that affect
the high bits (important because of signed/unsigned distinctions). Also,
included the C code in the comments so that that the reference data
may be more easily re-created.  Also, changed test to compare out to
14 digits.  Avoided the full 15, because some platforms may do weird
things to the last couple of bits.

Changed docstrings into comment blocks so they won't print-out during
the regression test.