Random.seed() now takes PyLong as well as PyInt.  Keeps interface consistent with the old random.py.
Put in a new .jumpahead(n) method.  The interface is consistent with
the old random.py, but there are new semantics.  Since a specific n jump
cannot be computed, takes a useful and compatible alternative by altering
the internal state to another state far from the original.  The new
semantics are faithful to the use case of supporting multiple threads
with non-overlapping sequences.

The MersenneTwister C code is now subclassable.  Accordingly, the random.Random
class now inherits from MersenneTwister.Random.