Brrrr.  Many changes to make comparison and subtraction of aware objects ignore tzinfo if the operands have identical tzinfo members (meaning object identity -- "is").  I misunderstood the intent here, reading a wrong conclusion into conflicting clues.
The Python implementation's datetime.__cmp__ now handles all comparisons
for datetime and datetimetz, and likewise for time.__cmp__ wrt timetz.
The C implementation does that already.  The problem was that, e.g.,
comparing a naive time to an aware timetz failed to raise TypeError,
as previously time.__cmp__ silently ignored that "other" was actually
of type timetz and just compared the common base fields.  Unfortunately,
no test case caught this, so added one.  Unfortunately again, the new
test has to be skipped when running the Python implementation under 2.2.2,
because it triggers a Python bug (which Guido already fixed in 2.2 CVS
for Python 2.2.3).