A few small fixes.
The blockoff must always be allocated based on the number of allocated
blocks, not the number of reachable blocks.  The old code used the
number of reachable blocks which failed when there were unreachable
blocks.  (Also free blockoff when we're done with it.)

Move compiler_unit_check() above the call to compiler_unit_free(),
then call check from free.

Initialize u_tmpname.