The format of the string data used in the imageop module is described as "This is the same format as used by gl.lrectwrite() and the imgfile module."  This implies a certain byte order in multi-byte pixel formats.  However, the code was originally written on an SGI (big-endian) and *uses* the fact that bytes are stored in a particular order in ints.  This means that the code uses and produces different byte order on little-endian systems.
This fix adds a module-level flag "backward_compatible" (default not
set, and if not set, behaves as if set to 1--i.e. backward compatible)
that can be used on a little-endian system to use the same byte order
as the SGI.  Using this flag it is then possible to prepare
SGI-compatible images on a little-endian system.

This patch is the result of a (small) discussion on python-dev and was
submitted to SourceForge as patch #874358.