Make DualObjTypeInfo more robust.
If the typeinfo cannot be loaded (typelib not registered, for
example), print a traceback, and continue.  All the IDispatch methods
will return E_NOTIMPL in this case, except that GetTypeInfoCount will
return 0.

First try on a addref_and_copy function, which copies a COM interface
pointer somewhere else, and calls AddRef() on it if it isn't NULL.