The setter methods return an object which must be kept alive, to keep the data valid which has been stored in the memory block.  The ctypes object instance inserts this object into its 'b_objects' list.
For simple Python types like integers or characters, there is nothing that
has to been kept alive, so Py_None is returned in these cases.  But this
makes inspecting the 'b_objects' list, which is accessible from Python for
debugging, less useful.

So, defining the _CTYPES_DEBUG_KEEP symbol returns the original value
instead of Py_None.