Changes merged in from the Stackless 2.4.3 branch..
Revision 47178: If your tasklet subclass was pickled, it would be reduced with the tasklet class rather than the subclass itself.  This meant when it was unpickled, it would be an instance of tasklet, rather than your subclass.  Now reduce uses the type of the tasklet, rather than being hardcoded to use the tasklet class itself.

Revision 47195: An example of using the watchdog in C.  Also implements a BeNice yielding function, but uses soft switching, so any tasklets which are blocked on it can be pickled.  Needs a Sleep function, and perhaps some more work.

Revision 47196: Changed the __reduce__ function for channels so that subclasses will get their class persisted and will be unpickled with the same custom class (rather than forced to have the tasklet class).

Revision 47240: Patch from Jeff Senn to fix OS X compilation problems.  One prevented Stackless from being compiled because one of the registers preserved was now clobbered in the generated assembly code.  The other was a warning because of the types used in slp_switch.