Re-implement the removal of the 'file' type's initializer.
file_init() was renamed _PyFile_Init() and made non-static.  At the C level, a
simulation of calling open() is now done with calling PyFile_Type.tp_new() and
passing the result to _PyFile_Init() with the tuple and dict arguments.

Exposed _PyFile_Init() as the file_init() function in the objcap module.  Takes
in an instance of the 'file' type and the typical arguments to open() and
initializes the instance as needed to refer to a file.  This allows subclasses
of 'file' to work with some changes.  __new__() will now need to be overridden
to make sure to not call file.__new__() with any arguments.  Then, in
__init__(), just pass 'self' to file_init() with the needed arguments to get
the open file bound to the instance (see Lib/test/test_descr.py for an example).
Doing this allows (eventual) import blocking of objcap and thus make open() the
only way to open new files if desired but still provide subclassing 'file' to
be functional if so desired.