Split the dispatcher into two parts.  New one is 'stacklesssocket'.
Why?  GC problems.  The old scheme required explict .close() calls
because asyncore's global socket_map kept a bound method from each
dispatcher.  This meant the dispatcher was never gc'ed if it was
no longer in use elsewhere.

The new scheme has stacklesssocket as a facade to dispatcher.
When the stacklesssocket gets a __del__ it forces the dispatcher
to close if not already closed.

Removed all explicit close statements from the test code.