Fix two problems.
Where we used to give a dispatcher back, now we give a socket facade.  So setting "wrap_accept_socket" then set it on the facade rather than the dispatcher, resulting in it not getting used.  Now setting this attribute on the facade forwards it on to the dispatcher.

Added a timeout to the asyncore polling, so that the socket module would not consume all the cpu.