New importer using the backup xml file from sourceforge instead of effbot's sourceforge scraper.
BeatifulSoup.py copied from the sourceforge scraper. 

Importer has support for most of the fields in the python-dev roundup
schema, with one notable exception in the status and resolution
handling, which need to be further researched given the large number
of different more or less invalid combinations of
status/resolution. See
http://psf.upfronthosting.co.za/roundup/meta/issue55 for details.