Collin Winter:
a) adds a get_lineno() method to pytree.py:Base. fix_except and
fix_raise were both using it, and I imagine future fixers will need
this functionality.

b) adds a cannot_convert() method to BaseFixer. This is a way for
fixers to indicate "this code won't work in Python 3, but I can't fix
it, and here's why". Both fix_except and fix_raise are modified to use
this.