Merged in r61151-62123 (equivalent to the r26a2 tag) from the mainline trunk.
'configure' is still the pre-2.6a1 version.  It still needs to be regenerated and checked in.

There are still deprecation warnings happening when you subclass the tasklet class and have it call the base '__init__' function with the passed in function as an argument when it gets instantiated.  This appears when running the Stackless unittests, in 'test_watchdog.py'.