Moved lib-tk to Lib/lib-tk so it matches the correct python structure.
Added Doc/library. tk.rst.diff is a minor diff for adding ttk.rst to
its toctree. ttk.rst is empty for now.

Removed unused import at Lib/lib-tk/Ttk.py and now it overrides
Tkinter.Tk._loadtk to check for Ttk presence. Previously it was
overriding Tkinter.BaseWidget._setup (which would call
Tkinter.Tk._loadtk) but in some situations it would be bad to
depend on BaseWidget, for example:

  import Tkinter, Ttk

  tk = Tkinter.Tk() # Tile not checked yet, not a problem for Tk 8.5
  style = Ttk.Style() # fails telling ttk::style is not available

The difference now is that it will raise a TclError when creating
Tk instance if you are trying to use Ttk without it being available.