method element_create of class Style is wrapped now.
method layout of class Style correctly returns the layout of an style
if a layoutspec is not passed, this was True before but I forgot about
this feature on my last commit and it was accidentally removed.

Added a sample demonstrating element_create and layout in short for
creating a rounded frame.