Fix os.kill tests to be more robust and work with slower machines.
Rather than depending on some sleep value, start up an interpreter
as a subprocess and communicate with it. Because subprocess pipes 
can't be read from until EOF and I want to read from them before that,
use ctypes to peek by using PeekNamedPipe. Once the subprocess has 
written the message, then it is ready to roll and accept signals.
After that, kill it.