Merged revisions 80800 via svnmerge from svn+ssh://pythondev@svn.python.org/python/trunk
It turns out that email5 (py3k), because it is using unicode for the
payload, doesn't do the encoding to the output character set until later
in the process.  Specifically, charset.body_encode no longer does the
input-to-output charset conversion.  So the if test in the exception
clause in encoders.encode_7or8bit really is needed in email5.

So, this merge only merges the test, not the removal of the 'if'.

........
  r80800 | r.david.murray | 2010-05-05 13:31:03 -0400 (Wed, 05 May 2010) | 9 lines
  
  Issue #7472: remove unused code from email.encoders.encode_7or8bit.
  
  Yukihiro Nakadaira noticed a typo in encode_7or8bit that was trying
  to special case iso-2022 codecs.  It turns out that the code in
  question is never used, because whereas it was designed to trigger
  if the payload encoding was eight bit but its output encoding was
  7 bit, in practice the payload is always converted to the 7bit 
  encoding before encode_7or8bit is called.  Patch by Shawat Anand.
........