Without this patch the value of sysconfig.get_config_var('LDSHARED') is wrong when PY_LDFLAGS is not empty.
The bug was caused by LDSHARED getting expanded *before* sysconfig
renamed PY_LDSHARED (and simular values) to names without a PY_
prefix.

The patch tries to maintain the intended behaviour of allowing users
to set LDFLAGS in the environment and have that affect the build.

Without this patch a universal build on OSX cannot build universal
(fat binary) extensions.